requirements.txt
boto3>=1.34.0
botocore>=1.34.0
openai>=1.40.0
anthropic>=0.34.0
google-generativeai>=0.8.3
lxml>=5.2.2
requests>=2.32.3
typer>=0.12.5
